requirements.txt
aniso8601==9.0.1
click==8.0.3
dnspython==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pymongo==3.12.1
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
Werkzeug==2.0.2
